cutler

Module commands

Functions

apply_defaults

Applies settings from the configuration file

check_for_updates

Checks for updates to cutler by fetching the latest version from GitHub’s Cargo.toml

config_delete

Deletes the configuration file and offers to unapply settings if they are still active

config_show

Displays the contents of the configuration file to the terminal.

init_config

Initializes a new cutler configuration file with sensible defaults.

reset_defaults

Hard resets all domains from the config file without using the snapshot

restart_system_services

Kills (restarts) Finder, Dock, and SystemUIServer to refresh settings

status_defaults

Displays the current status comparing the config vs current defaults

unapply_defaults

Unapplies settings using the stored snapshot